longer

+++ src/SalesManager.java
@@ -1,11 +1,11 @@
 public class SalesManager {
-    protected int[] sales;
-    public SalesManager(int[] sales) {
+    protected long[] sales;
+    public SalesManager(long[] sales) {
         this.sales = sales;
     }
-    public int max() {
-        int max = -1;
-        for (int sale : sales) {
+    public long max() {
+        long max = -1;
+        for (long sale : sales) {
             if (sale > max) {
                 max = sale;
             }
